Refactoring to remove _src private module (#40)
* refactoring to remove _src private module

* ignore cache directories

* missed import
