--- src/jaxoplanet/core.py
@@ -1,2 +1,0 @@
-from jaxoplanet._src.core.kepler import kepler as kepler
-from jaxoplanet._src.core.quad import light_curve as light_curve
