+++ src/jaxoplanet/core/__init__.py
@@ -1,0 +1,2 @@
+from jaxoplanet.core.kepler import kepler as kepler
+from jaxoplanet.core.quad import light_curve as light_curve
