+++ src/jaxoplanet/core/kepler.py
@@ -3,7 +3,7 @@ from typing import Tuple
 import jax
 import jax.numpy as jnp
 
-from jaxoplanet._src.types import Array
+from jaxoplanet.types import Array
 
 
 def kepler(M: Array, ecc: Array) -> Tuple[Array, Array]:
