+++ src/jaxoplanet/experimental/starry/__init__.py
@@ -1,0 +1,3 @@
+from jaxoplanet.experimental.starry.solution import (
+    solution_vector as solution_vector,
+)
