--- src/jaxoplanet/orbits.py
@@ -1,4 +1,0 @@
-from jaxoplanet._src.keplerian import KeplerianBody as KeplerianBody
-from jaxoplanet._src.keplerian import KeplerianCentral as KeplerianCentral
-from jaxoplanet._src.keplerian import KeplerianOrbit as KeplerianOrbit
-from jaxoplanet._src.transit import TransitOrbit as TransitOrbit
