+++ tests/core/quad_test.py
@@ -3,7 +3,7 @@
 import jax
 import jax.numpy as jnp
 import pytest
-from jaxoplanet._src.core.quad import quad_soln_impl
+from jaxoplanet.core.quad import quad_soln_impl
 from jaxoplanet.test_utils import assert_allclose
