+++ tests/experimental/starry/rotation_test.py
@@ -1,7 +1,7 @@
 import jax
 import numpy as np
 import pytest
-from jaxoplanet._src.experimental.starry.rotation import (
+from jaxoplanet.experimental.starry.rotation import (
     R_full,
     Rdot,
     axis_to_euler,
