+++ tests/experimental/starry/solution_test.py
@@ -5,7 +5,7 @@ import warnings
 import jax
 import numpy as np
 import pytest
-from jaxoplanet._src.experimental.starry.solution import kappas, solution_vector
+from jaxoplanet.experimental.starry.solution import kappas, solution_vector
 from jaxoplanet.test_utils import assert_allclose
